+++ src/fast16_score.h
@@ -62,7 +62,7 @@ namespace cv
 void makeOffsets(int pixel[25], int row_stride, int patternSize);
 
 template<int patternSize>
-int cornerScore(const uchar* ptr, const int pixel[], int threshold);
+int cornerScore(const ushort* ptr, const int pixel[], int threshold);
 
 }
